finish raw testcase
