+++ ut_mem_block/load_queue/UncacheBuffer/util/dataclass.py
@@ -1,0 +1,161 @@
+__all__ = [name for name in locals()]
+
+class IORedirect:
+    valid = False
+    robIdx_flag = False
+    # 8 bits
+    robIdx_value = 0
+    level = False
+    
+class RobIO:
+    pendingMMIOld = False
+    pendingPtr_flag = False
+    pendingPtr_value = 0  # 8-bit
+    mmio_0 = False   
+    mmio_1 = False   
+    mmio_2 = False
+    uop_0_robIdx_value = 0  # 8-bit
+    uop_1_robIdx_value = 0  # 8-bit
+    uop_2_robIdx_value = 0  # 8-bit
+    
+class ReqIO:
+    valid = False
+    bits_uop_exceptionVec_3 = False
+    bits_uop_exceptionVec_4 = False
+    bits_uop_exceptionVec_5 = False
+    bits_uop_exceptionVec_13 = False
+    bits_uop_exceptionVec_21 = False
+    bits_uop_trigger = 0  # 4-bit value
+    bits_uop_preDecodeInfo_isRVC = False
+    bits_uop_ftqPtr_flag = False
+    bits_uop_ftqPtr_value = 0  # 6-bit value
+    bits_uop_ftqOffset = 0  # 4-bit value
+    bits_uop_fuOpType = 0  # 9-bit value
+    bits_uop_rfWen = False
+    bits_uop_fpWen = False
+    bits_uop_vpu_vstart = 0  # 8-bit value
+    bits_uop_vpu_veew = 0  # 2-bit value
+    bits_uop_uopIdx = 0  # 7-bit value
+    bits_uop_pdest = 0  # 8-bit value
+    bits_uop_robIdx_flag = False
+    bits_uop_robIdx_value = 0  # 8-bit value
+    bits_uop_storeSetHit = False
+    bits_uop_waitForRobIdx_flag = False
+    bits_uop_waitForRobIdx_value = 0  # 8-bit value
+    bits_uop_loadWaitBit = False
+    bits_uop_loadWaitStrict = False
+    bits_uop_lqIdx_flag = False
+    bits_uop_lqIdx_value = 0  # 7-bit value
+    bits_uop_sqIdx_flag = False
+    bits_uop_sqIdx_value = 0  # 6-bit value
+    bits_vaddr = 0  # 50-bit value
+    bits_fullva = 0  # 64-bit value
+    bits_isHyper = False
+    bits_paddr = 0  # 48-bit value
+    bits_gpaddr = 0  # 64-bit value
+    bits_isForVSnonLeafPTE = False
+    bits_mask = 0  # 16-bit value
+    bits_nc = False
+    bits_mmio = False
+    bits_memBackTypeMM = False
+    bits_isvec = False
+    bits_is128bit = False
+    bits_vecActive = False
+    bits_schedIndex = 0  # 7-bit value
+    bits_rep_info_cause_0 = False
+    bits_rep_info_cause_1 = False
+    bits_rep_info_cause_2 = False
+    bits_rep_info_cause_3 = False
+    bits_rep_info_cause_4 = False
+    bits_rep_info_cause_5 = False
+    bits_rep_info_cause_6 = False
+    bits_rep_info_cause_7 = False
+    bits_rep_info_cause_8 = False
+    bits_rep_info_cause_9 = False
+    bits_rep_info_cause_10 = False
+    
+class MmioOut:
+    ready = False
+    valid = False
+    bits_uop_exceptionVec_3 = False
+    bits_uop_exceptionVec_4 = False
+    bits_uop_exceptionVec_5 = False
+    bits_uop_exceptionVec_13 = False
+    bits_uop_exceptionVec_19 = False
+    bits_uop_exceptionVec_21 = False
+    bits_uop_trigger = 0  # 4-bit value
+    bits_uop_preDecodeInfo_isRVC = False
+    bits_uop_ftqPtr_flag = False
+    bits_uop_ftqPtr_value = 0  # 6-bit value
+    bits_uop_ftqOffset = 0  # 4-bit value
+    bits_uop_fuOpType = 0  # 9-bit value
+    bits_uop_rfWen = False
+    bits_uop_fpWen = False
+    bits_uop_flushPipe = False
+    bits_uop_vpu_vstart = 0  # 8-bit value
+    bits_uop_vpu_veew = 0  # 2-bit value
+    bits_uop_uopIdx = 0  # 7-bit value
+    bits_uop_pdest = 0  # 8-bit value
+    bits_uop_robIdx_flag = False
+    bits_uop_robIdx_value = 0  # 8-bit value
+    bits_uop_storeSetHit = False
+    bits_uop_waitForRobIdx_flag = False
+    bits_uop_waitForRobIdx_value = 0  # 8-bit value
+    bits_uop_loadWaitBit = False
+    bits_uop_loadWaitStrict = False
+    bits_uop_lqIdx_flag = False
+    bits_uop_lqIdx_value = 0  # 7-bit value
+    bits_uop_sqIdx_flag = False
+    bits_uop_sqIdx_value = 0  # 6-bit value
+    bits_uop_replayInst = False
+
+class NcOut:
+    ready = False
+    valid = False
+    bits_uop_exceptionVec_4 = False
+    bits_uop_exceptionVec_19 = False
+    bits_uop_preDecodeInfo_isRVC = False
+    bits_uop_ftqPtr_flag = False
+    bits_uop_ftqPtr_value = 0  # 6-bit value
+    bits_uop_ftqOffset = 0  # 4-bit value
+    bits_uop_fuOpType = 0  # 9-bit value
+    bits_uop_rfWen = False
+    bits_uop_fpWen = False
+    bits_uop_vpu_vstart = 0  # 8-bit value
+    bits_uop_vpu_veew = 0  # 2-bit value
+    bits_uop_uopIdx = 0  # 7-bit value
+    bits_uop_pdest = 0  # 8-bit value
+    bits_uop_robIdx_flag = False
+    bits_uop_robIdx_value = 0  # 8-bit value
+    bits_uop_storeSetHit = False
+    bits_uop_waitForRobIdx_flag = False
+    bits_uop_waitForRobIdx_value = 0  # 8-bit value
+    bits_uop_loadWaitBit = False
+    bits_uop_loadWaitStrict = False
+    bits_uop_lqIdx_flag = False
+    bits_uop_lqIdx_value = 0  # 7-bit value
+    bits_uop_sqIdx_flag = False
+    bits_uop_sqIdx_value = 0  # 6-bit value
+    bits_vaddr = 0  # 50-bit value
+    bits_paddr = 0  # 48-bit value
+    bits_data = 0  # 128-bit value
+    bits_isvec = False
+    bits_is128bit = False
+    bits_vecActive = False
+    bits_schedIndex = 0  # 7-bit value
+    
+class Uncache:
+    req_ready = False
+    req_valid = False
+    req_bits_cmd = 0  # 5-bit value
+    req_bits_addr = 0  # 48-bit value
+    req_bits_vaddr = 0  # 50-bit value
+    req_bits_data = 0  # 64-bit value
+    req_bits_mask = 0  # 8-bit value
+    req_bits_id = 0  # 7-bit value
+    req_bits_nc = False
+    req_bits_memBackTypeMM = False
+    resp_valid = False
+    resp_bits_data = 0  # 64-bit value
+    resp_bits_id = 0  # 7-bit value
+    resp_bits_nderr = False
